Split raw_map into a new street_network crate.
MapConfig has a new bool, so we will need to regenerate all data.
