+++ convert_osm/src/split_ways.rs
@@ -56,7 +56,7 @@ pub fn split_up_roads(map: &mut RawMap, mut input: OsmExtract, timer: &mut Timer
     }
 
     for (pt, id) in &pt_to_intersection {
-        map.intersections.insert(
+        map.streets.intersections.insert(
             *id,
             RawIntersection::new(
                 pt.to_pt2d(),
@@ -71,7 +71,8 @@ pub fn split_up_roads(map: &mut RawMap, mut input: OsmExtract, timer: &mut Timer
 
     // Set roundabouts to their center
     for (id, point) in roundabout_centers {
-        map.intersections
+        map.streets
+            .intersections
             .insert(id, RawIntersection::new(point, IntersectionType::StopSign));
     }
 
@@ -113,9 +114,9 @@ pub fn split_up_roads(map: &mut RawMap, mut input: OsmExtract, timer: &mut Timer
                 }
 
                 let osm_center_pts = simplify_linestring(std::mem::take(&mut pts));
-                match RawRoad::new(osm_center_pts, tags, &map.config) {
+                match RawRoad::new(osm_center_pts, tags, &map.streets.config) {
                     Ok(road) => {
-                        map.roads.insert(id, road);
+                        map.streets.roads.insert(id, road);
                     }
                     Err(err) => {
                         error!("Skipping {id}: {err}");
@@ -136,7 +137,7 @@ pub fn split_up_roads(map: &mut RawMap, mut input: OsmExtract, timer: &mut Timer
     // Resolve simple turn restrictions (via a node)
     let mut restrictions = Vec::new();
     for (restriction, from_osm, via_osm, to_osm) in input.simple_turn_restrictions {
-        let roads = map.roads_per_intersection(via_osm);
+        let roads = map.streets.roads_per_intersection(via_osm);
         // If some of the roads are missing, they were likely filtered out -- usually service
         // roads.
         if let (Some(from), Some(to)) = (
@@ -147,7 +148,8 @@ pub fn split_up_roads(map: &mut RawMap, mut input: OsmExtract, timer: &mut Timer
         }
     }
     for (from, rt, to) in restrictions {
-        map.roads
+        map.streets
+            .roads
             .get_mut(&from)
             .unwrap()
             .turn_restrictions
@@ -159,6 +161,7 @@ pub fn split_up_roads(map: &mut RawMap, mut input: OsmExtract, timer: &mut Timer
     let mut complicated_restrictions = Vec::new();
     for (rel_osm, from_osm, via_osm, to_osm) in input.complicated_turn_restrictions {
         let via_candidates: Vec<OriginalRoad> = map
+            .streets
             .roads
             .keys()
             .filter(|r| r.osm_way_id == via_osm)
@@ -175,14 +178,16 @@ pub fn split_up_roads(map: &mut RawMap, mut input: OsmExtract, timer: &mut Timer
         let via = via_candidates[0];
 
         let maybe_from = map
+            .streets
             .roads_per_intersection(via.i1)
             .into_iter()
-            .chain(map.roads_per_intersection(via.i2).into_iter())
+            .chain(map.streets.roads_per_intersection(via.i2).into_iter())
             .find(|r| r.osm_way_id == from_osm);
         let maybe_to = map
+            .streets
             .roads_per_intersection(via.i1)
             .into_iter()
-            .chain(map.roads_per_intersection(via.i2).into_iter())
+            .chain(map.streets.roads_per_intersection(via.i2).into_iter())
             .find(|r| r.osm_way_id == to_osm);
         match (maybe_from, maybe_to) {
             (Some(from), Some(to)) => {
@@ -197,7 +202,8 @@ pub fn split_up_roads(map: &mut RawMap, mut input: OsmExtract, timer: &mut Timer
         }
     }
     for (from, via, to) in complicated_restrictions {
-        map.roads
+        map.streets
+            .roads
             .get_mut(&from)
             .unwrap()
             .complicated_turn_restrictions
@@ -210,10 +216,16 @@ pub fn split_up_roads(map: &mut RawMap, mut input: OsmExtract, timer: &mut Timer
     for (pt, dir) in input.traffic_signals {
         if let Some(r) = pt_to_road.get(&pt) {
             // Example: https://www.openstreetmap.org/node/26734224
-            if !map.roads[r].osm_tags.is(osm::HIGHWAY, "construction") {
+            if !map.streets.roads[r]
+                .osm_tags
+                .is(osm::HIGHWAY, "construction")
+            {
                 let i = if dir == Direction::Fwd { r.i2 } else { r.i1 };
-                map.intersections.get_mut(&i).unwrap().intersection_type =
-                    IntersectionType::TrafficSignal;
+                map.streets
+                    .intersections
+                    .get_mut(&i)
+                    .unwrap()
+                    .intersection_type = IntersectionType::TrafficSignal;
             }
         }
     }
