+++ street_network/Cargo.toml
@@ -1,14 +1,18 @@
 [package]
-name = "raw_map"
+name = "street_network"
 version = "0.1.0"
 authors = ["Dustin Carlino <dabreegster@gmail.com>"]
 edition = "2021"
 
 [dependencies]
+aabb-quadtree = "0.1.0"
 abstio = { path = "../abstio" }
 abstutil = { path = "../abstutil" }
+anyhow = "1.0.38"
+geojson = { version = "0.22.2", features = ["geo-types"] }
 geom = { path = "../geom" }
+kml = { path = "../kml" }
+log = "0.4.14"
+petgraph = { version = "0.6.0" }
 serde = "1.0.123"
-street_network = { path = "../street_network" }
-strum = "0.21"
-strum_macros = "0.21"
+serde_json = "1.0.61"
