+++ street_network/src/geometry/algorithm.rs
